+++ source/vst/aaxwrapper/aaxwrapper_parameters.cpp
@@ -717,7 +717,7 @@ AAX_Result AAXWrapper_Parameters::GetChunk (AAX_CTypeID chunkID, AAX_SPlugInChun
 	FGuard guard (mWrapper->mSyncCalls);
 	// assume GetChunkSize called before and size of oChunk correct
 	oChunk->fVersion = 1;
-	oChunk->fSize = mWrapper->mChunk.getSize ();
+	oChunk->fSize = static_cast<int32_t> (mWrapper->mChunk.getSize ());
 	memcpy (oChunk->fData, mWrapper->mChunk.getData (),
 	        static_cast<size_t> (mWrapper->mChunk.getSize ()));
 	strncpy (reinterpret_cast<char*> (oChunk->fName), AAXWRAPPER_CONTROLS_CHUNK_DESCRIPTION, 31);
